chore: remove testing type specific servers
