+++ packages/server/test/integration/cypress_spec.js
@@ -31,7 +31,7 @@ const cache = require(`../../lib/cache`)
 const errors = require(`../../lib/errors`)
 const cypress = require(`../../lib/cypress`)
 const ProjectBase = require(`../../lib/project-base`).ProjectBase
-const { ServerE2E } = require(`../../lib/server-e2e`)
+const { ServerBase } = require(`../../lib/server-base`)
 const Reporter = require(`../../lib/reporter`)
 const browsers = require(`../../lib/browsers`)
 const videoCapture = require(`../../lib/video_capture`)
@@ -174,7 +174,7 @@ describe('lib/cypress', () => {
     sinon.stub(extension, 'setHostAndPath').resolves()
     sinon.stub(detect, 'detect').resolves(TYPICAL_BROWSERS)
     sinon.stub(process, 'exit')
-    sinon.stub(ServerE2E.prototype, 'reset')
+    sinon.stub(ServerBase.prototype, 'reset')
     sinon.stub(errors, 'warning')
     .callThrough()
     .withArgs('INVOKED_BINARY_OUTSIDE_NPM_MODULE')
@@ -1102,7 +1102,7 @@ describe('lib/cypress', () => {
 
       it('can change the default port to 5544', function () {
         const listen = sinon.spy(http.Server.prototype, 'listen')
-        const open = sinon.spy(ServerE2E.prototype, 'open')
+        const open = sinon.spy(ServerBase.prototype, 'open')
 
         return cypress.start([`--run-project=${this.todosPath}`, '--port=5544'])
         .then(() => {
@@ -1777,7 +1777,7 @@ describe('lib/cypress', () => {
 
       sinon.stub(electron.app, 'on').withArgs('ready').yieldsAsync()
       sinon.stub(Windows, 'open').resolves(this.win)
-      sinon.stub(ServerE2E.prototype, 'startWebsockets')
+      sinon.stub(ServerBase.prototype, 'startWebsockets')
       sinon.stub(electron.ipcMain, 'on')
     })
 
@@ -1798,7 +1798,7 @@ describe('lib/cypress', () => {
 
     // TODO: fix failing test https://github.com/cypress-io/cypress/issues/23149
     it.skip('passes filtered options to Project#open and sets cli config', async function () {
-      const open = sinon.stub(ServerE2E.prototype, 'open').resolves([])
+      const open = sinon.stub(ServerBase.prototype, 'open').resolves([])
 
       sinon.stub(interactiveMode, 'ready')
