+++ packages/server/test/integration/websockets_spec.js
@@ -8,7 +8,7 @@ const Promise = require('bluebird')
 const socketIo = require(`@packages/socket/lib/browser`)
 const httpsServer = require(`@packages/https-proxy/test/helpers/https_server`)
 const config = require(`../../lib/config`)
-const { ServerE2E } = require(`../../lib/server-e2e`)
+const { ServerBase } = require(`../../lib/server-base`)
 const { SocketE2E } = require(`../../lib/socket-e2e`)
 const { Automation } = require(`../../lib/automation`)
 const Fixtures = require('@tooling/system-tests')
@@ -38,7 +38,7 @@ describe('Web Sockets', () => {
       this.cfg = cfg
       this.ws = new ws.Server({ port: wsPort })
 
-      this.server = new ServerE2E()
+      this.server = new ServerBase()
 
       return this.server.open(this.cfg, {
         SocketCtor: SocketE2E,
