+++ packages/server/test/performance/proxy_performance_spec.js
@@ -22,7 +22,7 @@ process.env.CYPRESS_INTERNAL_ENV = 'development'
 
 const CA = require('@packages/https-proxy').CA
 const { setupFullConfigWithDefaults } = require('@packages/config')
-const { ServerE2E } = require('../../lib/server-e2e')
+const { ServerBase } = require('../../lib/server-base')
 const { SocketE2E } = require('../../lib/socket-e2e')
 const { _getArgs } = require('../../lib/browsers/chrome')
 
@@ -361,7 +361,7 @@ describe('Proxy Performance', function () {
           // turn off morgan
           config.morgan = false
 
-          cyServer = new ServerE2E()
+          cyServer = new ServerBase()
 
           return cyServer.open(config, {
             SocketCtor: SocketE2E,
